Update the cadnano->oxDNA test
Note that the Python3 version of the cadnano->oxDNA script works better
than the Python2 version. Indeed, the latter sometimes generated bonded
nucleotides that were at slightly off angles with respect to the
"optimal value". I did not manage to understand where and when the issue
appeared in the codebase, but it is most likely due to the different way
that integer division is handled in python2 vs python3.