Fix a bug in the cadnano converter
The converter would fail when passed json files containing some virtual
helices with a scaffold and some without. In general, the behaviour in
these cases is undefined since staples are not converted if there is no
scaffold. However, sometimes vhelices would be completely devoid of
scaffolds AND staples and in these cases we want to keep on converting
anyway. The user is warned whenever scaffold-less vhelices are
encountered.